SQLProvider on comments update
