+++ .gitignore
@@ -3,3 +3,4 @@ src/PrizmMainProject/Properties/licenses.licx
 *.bak
 src/PrizmMainProject/Properties/licenses.licx
 [Ee]xternal/
+src/PrizmMainProject/App.config
