+++ src/Data/Data.csproj
@@ -94,6 +94,7 @@
     <Compile Include="DAL\IMillReportsRepository.cs" />
     <Compile Include="DAL\IProjectRepository.cs" />
     <Compile Include="DAL\IRepository.cs" />
+    <Compile Include="DAL\ISQLFlexible.cs" />
     <Compile Include="DAL\Mapping\AuditLogMap.cs" />
     <Compile Include="DAL\Mapping\CategoryMap.cs" />
     <Compile Include="DAL\Mapping\CoatMap.cs" />
