+++ src/PrizmMainProject/App.config
@@ -11,7 +11,7 @@
   </startup>
 
   <connectionStrings>
-    <add name="PrizmDatabase" connectionString="Data Source=(LocalDb)\v11.0;Initial Catalog=prizm;Integrated Security=true;AttachDBFileName=C:\Program Files\Prizm\prj20141221\Data\prj20141221.mdf" />
+    <add name="PrizmDatabase" connectionString="Data Source=(LocalDb)\v11.0;Initial Catalog=prizm;AttachDBFileName=C:\Program Files\Prizm\prj20141221\Data\prj20141221.mdf;Integrated Security=true" />
   </connectionStrings>
 
   <appSettings>
